+++ agent-control/tests/on_host/command/shutdown.rs
@@ -1,8 +1,8 @@
 use newrelic_agent_control::sub_agent::on_host::command::command_os::CommandOSNotStarted;
 use newrelic_agent_control::sub_agent::on_host::command::executable_data::ExecutableData;
+use newrelic_agent_control::sub_agent::on_host::command::logging::file_logger::FileLoggingConfigSubAgent;
 use newrelic_agent_control::sub_agent::on_host::command::restart_policy::RestartPolicy;
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::process::Command;
 use std::{thread::sleep, time::Duration};
 
@@ -32,8 +32,7 @@ fn non_blocking_runner() {
             restart_policy: RestartPolicy::default(),
             shutdown_timeout: Default::default(),
         },
-        false,
-        PathBuf::default(),
+        FileLoggingConfigSubAgent::default(),
     );
 
     let mut started_cmd = cmd.start().unwrap();
@@ -57,8 +56,7 @@ fn command_shutdown_when_sigterm_is_ignored() {
             restart_policy: Default::default(),
             shutdown_timeout: Duration::from_millis(100),
         },
-        false,
-        Default::default(),
+        FileLoggingConfigSubAgent::default(),
     )
     .start()
     .unwrap();
@@ -108,8 +106,7 @@ fn command_shutdown_kill_orphan_process() {
             restart_policy: Default::default(),
             shutdown_timeout: Duration::from_millis(100),
         },
-        false,
-        Default::default(),
+        FileLoggingConfigSubAgent::default(),
     )
     .start()
     .unwrap();